Convert Runs to TS
