+++ client/components/Runs/Runs.story.tsx
@@ -1,0 +1,25 @@
+import { storiesOf } from '@storybook/react';
+import React from 'react';
+import Runs from './';
+
+storiesOf('Runs', module).add('default', () => (
+  <Runs
+    {...{
+      name: 'Export',
+      status: 'idle',
+      runs: [
+        {
+          ID: '1',
+          status: 'finished',
+          scheduled_at: 'Yesterday',
+          started_at: 'Today',
+          finished_at: 'Tomorrow',
+          job: 'export',
+          rest_url: '',
+          job_url: '',
+          console_url: ''
+        }
+      ]
+    }}
+  />
+));
